+++ RELEASE_NOTES.md
@@ -84,6 +84,7 @@ Apache Royale Compiler 1.0.0
 - compiler: Fixed rare race condition where certain cached values are detected as non-null but become null before the method returns.
 - compiler: Fixed parsing of namespace uri and manifest mappings to allow multiple manifest paths per URI.
 - compiler: Fixed missing require for package-level functions or variables in some situations.
+- compiler: Fixed missing binding data caused by thread conflicts.
 - debugger: Added missing isolate ID to SWF load and unload events.
 - debugger: Fixed debugger targeting the current JDK version instead of the intended minimum JDK version.
 - debugger: Fixed localized messages appearing as unprocessed tokens.